requirements.txt
absl-py==2.2.0
ale-py==0.10.2
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.12
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
gymnasium==1.1.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
maturin==1.8.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
sentry-sdk==2.24.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
stable_baselines3 @ git+https://github.com/DLR-RM/stable-baselines3@656de97269c9e3051d3bbfb3f5f328d486867bd8
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tinyscaler==1.2.8
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.8
Werkzeug==3.1.3
zipp==3.21.0
